Move job info modal to own component file
