+++ src/components/Jobs/JobInfoList.jsx
@@ -1,11 +1,9 @@
-import { Modal } from 'antd';
-import axios from 'axios';
 import { useSelector } from 'react-redux';
 import { useEffect, useRef, useState } from 'react';
 
 import './JobInfoList.css';
-import socket from '../../socket';
 import JobInfoCard from './JobInfoCard';
+import JobInfoModal from './JobInfoModal';
 
 export const JobInfoList = () => {
   const jobs = useSelector((state) => state.jobs);
@@ -74,76 +72,13 @@ export const JobInfoList = () => {
         </div>
       )}
       {selectedJob && (
-        <JobOutputModal
+        <JobInfoModal
           job={selectedJob}
           visible={modalVisible}
           setVisible={setModalVisible}
         />
       )}
     </>
-  );
-};
-
-const JobOutputModal = ({ job, visible, setVisible }) => {
-  const [message, setMessage] = useState('');
-  const isFirst = useRef(true);
-  const listenerFuncRef = useRef(null);
-
-  const message_appender = (data) => {
-    if (data.jobid == job.id) {
-      setMessage((message) => message.concat(data.message));
-    }
-  };
-
-  useEffect(() => {
-    const getJobOutput = async () => {
-      try {
-        const resp = await axios.get(
-          `${import.meta.env.VITE_CEA_URL}/server/streams/read/${job.id}`,
-          null,
-          { responseType: 'text' },
-        );
-        setMessage(resp.data);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-    listenerFuncRef.current &&
-      socket.off('cea-worker-message', listenerFuncRef.current);
-    isFirst.current = true;
-    getJobOutput();
-  }, [job]);
-
-  useEffect(() => {
-    if (isFirst.current) {
-      listenerFuncRef.current = message_appender;
-      socket.on('cea-worker-message', message_appender);
-      isFirst.current = false;
-    }
-  }, [message]);
-
-  return (
-    <Modal
-      title={`Job Output for ${job.id} - ${job.script}`}
-      open={visible}
-      width={800}
-      footer={false}
-      onCancel={() => setVisible(false)}
-      destroyOnClose
-    >
-      <div style={{ height: '35vh' }}>
-        <pre
-          style={{
-            height: '90%',
-            overflow: 'auto',
-            fontSize: 12,
-            whiteSpace: 'pre-wrap',
-          }}
-        >
-          {message}
-        </pre>
-      </div>
-    </Modal>
   );
 };
