+++ src/components/Jobs/JobInfoModal.jsx
@@ -1,0 +1,70 @@
+import { Modal } from 'antd';
+import axios from 'axios';
+import { useEffect, useRef, useState } from 'react';
+
+import socket from '../../socket';
+
+const JobOutputModal = ({ job, visible, setVisible }) => {
+  const [message, setMessage] = useState('');
+  const isFirst = useRef(true);
+  const listenerFuncRef = useRef(null);
+
+  const message_appender = (data) => {
+    if (data.jobid == job.id) {
+      setMessage((message) => message.concat(data.message));
+    }
+  };
+
+  useEffect(() => {
+    const getJobOutput = async () => {
+      try {
+        const resp = await axios.get(
+          `${import.meta.env.VITE_CEA_URL}/server/streams/read/${job.id}`,
+          null,
+          { responseType: 'text' },
+        );
+        setMessage(resp.data);
+      } catch (error) {
+        console.error(error);
+      }
+    };
+    listenerFuncRef.current &&
+      socket.off('cea-worker-message', listenerFuncRef.current);
+    isFirst.current = true;
+    getJobOutput();
+  }, [job]);
+
+  useEffect(() => {
+    if (isFirst.current) {
+      listenerFuncRef.current = message_appender;
+      socket.on('cea-worker-message', message_appender);
+      isFirst.current = false;
+    }
+  }, [message]);
+
+  return (
+    <Modal
+      title={`Job Output for ${job.id} - ${job.script}`}
+      open={visible}
+      width={800}
+      footer={false}
+      onCancel={() => setVisible(false)}
+      destroyOnClose
+    >
+      <div style={{ height: '35vh' }}>
+        <pre
+          style={{
+            height: '90%',
+            overflow: 'auto',
+            fontSize: 12,
+            whiteSpace: 'pre-wrap',
+          }}
+        >
+          {message}
+        </pre>
+      </div>
+    </Modal>
+  );
+};
+
+export default JobOutputModal;
